Cast types of expected values in unit tests to be 64 bit compatible

+++ CorePitch/CorePitchTests/CPPitchManagerAudioQueueTests.m
@@ -44,7 +44,7 @@
     
     XCTAssertEqual(self.pitchManager.inputStreamDescription.mFramesPerPacket, (UInt32)1);
     XCTAssertEqual(self.pitchManager.inputStreamDescription.mChannelsPerFrame, (UInt32)1);
-    XCTAssertEqual(self.pitchManager.inputStreamDescription.mBitsPerChannel, sizeof(float) * 8);
+    XCTAssertEqual(self.pitchManager.inputStreamDescription.mBitsPerChannel, (UInt32)(sizeof(float) * 8));
 }
 
 - (void)testAudioQueue
